image metadata to cloudsearch #43, timestamp of ingest time to item, celery tweaking, timeout to image url open

+++ embed/app/ingest.py
@@ -1,11 +1,12 @@
 import os
-import urllib
+import urllib2
 import math
 import subprocess
 import time
 import random
 import re
 import hashlib
+from datetime import datetime
 
 import simplejson as json
 import redis
@@ -27,6 +28,7 @@ identify_output_regular = re.compile(r'''
 S3_CHUNK_SIZE = int(os.getenv('S3_CHUNK_SIZE', 52428800))
 S3_DEFAULT_FOLDER = os.getenv('S3_DEFAULT_FOLDER', '')
 MAX_TASK_REPEAT = int(os.getenv('MAX_TASK_REPEAT', 1))
+URL_OPEN_TIMEOUT = int(os.getenv('URL_OPEN_TIMEOUT', 10))
 
 
 @task_queue.task
@@ -42,10 +44,14 @@ def ingestQueue(batch_id, item_id, task_id):
 		if task.type == 'del':
 			try:
 				item = Item(item_id)
-				filename = item.image_meta[task.url]['filename']
-			
-				if filename:
-					bucket.delete_key(S3_DEFAULT_FOLDER + filename)
+				
+				if task.url_order > 0:
+					filename = '%s/%s.jp2' % (item_id, task.url_order)
+
+				else:
+					filename = '%s.jp2' % item_id
+
+				bucket.delete_key(S3_DEFAULT_FOLDER + filename)
 			except NoItemInDb:
 				pass
 
@@ -62,7 +68,10 @@ def ingestQueue(batch_id, item_id, task_id):
 				filename = '/tmp/%s' % item_id
 				destination = '%s.jp2' % item_id
 			
-			urllib.urlretrieve (task.url, filename)
+			r = urllib2.urlopen(task.url, timeout=URL_OPEN_TIMEOUT)
+			f = open(filename, 'wb')
+			f.write(r.read())
+			f.close()
 		
 			if subprocess.check_output(['identify', '-format', '%m', filename]) != 'TIFF':
 				subprocess.call(['convert', '-compress', 'none', filename, '%s.tif' % filename])
@@ -74,8 +83,6 @@ def ingestQueue(batch_id, item_id, task_id):
 		
 			if test:
 				task.image_meta = json.loads(test.group('size_json'))
-				task.image_meta['filename'] = destination
-				task.image_meta['order'] = task.url_order
 			else:
 				raise Exception
 		
@@ -99,20 +106,7 @@ def ingestQueue(batch_id, item_id, task_id):
 			os.remove('%s.tif' % filename)
 
 			task.status = 'ok'
-		
-		# is this the task with highest id for the specific item? (only last created task for specific item
-		# has its data)
-		if task.item_data:
-			cloudsearch = getCloudSearch()
-			
-			if task.item_data.has_key('status') and task.item_data['status'] == 'deleted':
-				cloudsearch.delete(hashlib.sha512(item_id).hexdigest()[:128])
-			else:
-				item = Item(item_id, task.item_data)
-				cloudsearch.add(hashlib.sha512(item_id).hexdigest()[:128], {'id': item.id, 'title': item.title, 'creator': item.creator, 'source': item.source, 'institution': item.institution, 'institution_link': item.institution_link, 'license': item.license, 'description': item.description})
-			
-			cloudsearch.commit()
-		
+				
 		task.save()
 
 	except:
@@ -126,10 +120,10 @@ def ingestQueue(batch_id, item_id, task_id):
 		else:
 			task.status = 'error'
 			task.save()
-
+	
 	if task.increment_finished_item_tasks() >= task.item_tasks_count:
 		finalizeItem(batch_id, item_id, task.item_tasks_count)
-		
+	
 	return
 
 
@@ -141,6 +135,7 @@ def finalizeItem(batch_id, item_id, item_tasks_count):
 	
 	# the task with highest id for the specific item has all item data
 	item_data = item_tasks[-1].item_data
+	item_data['timestamp'] = datetime.utcnow().isoformat("T") + "Z"
 	
 	try:
 		old_item = Item(item_id)
@@ -149,7 +144,28 @@ def finalizeItem(batch_id, item_id, item_tasks_count):
 
 	if old_item:
 		if item_data.has_key('status') and item_data['status'] == 'deleted':
-			old_item.delete()
+			i = 0
+			
+			while MAX_TASK_REPEAT > i:
+				try:
+					cloudsearch = getCloudSearch()
+					cloudsearch.delete(hashlib.sha512(item_id).hexdigest()[:128])
+					cloudsearch.commit()
+					break
+				except:
+					if i < MAX_TASK_REPEAT:
+						rand = task.attempts + random.randint(task.attempts, task.attempts * 2)
+						time.sleep(rand)
+					
+					i += 1
+					
+					continue
+			
+			if MAX_TASK_REPEAT > i:
+				old_item.delete()
+			else:
+				item_tasks[-1].status = 'error'
+				item_tasks[-1].save()
 			
 			return
 		else:
@@ -162,6 +178,9 @@ def finalizeItem(batch_id, item_id, item_tasks_count):
 	for task in item_tasks:
 		if task.status == 'pending' or task.status == 'error':
 			error = True
+		# modification tasks never changes image_meta
+		elif task.type == 'mod':
+			pass
 		elif task.status == 'deleted':
 			# if the image is being realy deleted not only being reshaffled
 			if not task.url in item_data['url']:
@@ -170,20 +189,55 @@ def finalizeItem(batch_id, item_id, item_tasks_count):
 			item_data['image_meta'][task.url] = task.image_meta
 
 	if not error:
-		item = Item(item_id, item_data)	
-		item.save()
+		item = Item(item_id, item_data)
+		i = 0
+		ordered_image_meta = []
+		
+		for url in item.url:
+			tmp = item.image_meta[url]
+			tmp['url'] = url
+			ordered_image_meta.append(tmp)
+			
+		while MAX_TASK_REPEAT > i:
+			try:
+				cloudsearch = getCloudSearch()
+				cloudsearch.add(hashlib.sha512(item_id).hexdigest()[:128], {'id': item.id, 'title': item.title, 'creator': item.creator, 'source': item.source, 'institution': item.institution, 'institution_link': item.institution_link, 'license': item.license, 'description': item.description, 'url': json.dumps(item.url), 'timestamp': item.timestamp, 'image_meta': json.dumps(ordered_image_meta)})
+				cloudsearch.commit()
+				break
+			except:
+				if i < MAX_TASK_REPEAT:
+					rand = task.attempts + random.randint(task.attempts, task.attempts * 2)
+					time.sleep(rand)
+					
+				i += 1
+				continue
+			
+		if MAX_TASK_REPEAT > i:
+			item.save()
+		else:
+			item_tasks[-1].status = 'error'
+			item_tasks[-1].save()
+			cleanErrItem(item_id, len(item_data['image_meta']))
+
 	else:
-		cleanErrItem(item_id, item_data['image_meta'])
+		cleanErrItem(item_id, len(item_data['image_meta']))
 	
 	return
 
 
-def cleanErrItem(item_id, urls):
+def cleanErrItem(item_id, count):
 	try:
 		bucket = getBucket()
+		i = 0
 		
-		for url in urls.keys():
-			filename = urls[url]['filename']
+		while count > i:
+			if i == 0:
+				filename = unique_id
+			else:
+				filename = '%s/%s' % (unique_id, i)
+			
+			i += 1
+			
 			bucket.delete_key(S3_DEFAULT_FOLDER + filename)
 	except:
 		pass
